refactored registration form
